pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scriberr-gateway"
version = "0.1.0"
description = "Gateway service to upload files to Scriberr and send Apprise notifications"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Your Name", email = "you@example.com"}
]
keywords = ["scriberr", "apprise", "fastapi", "gateway"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "requests>=2.31.0",
  "PyYAML>=6.0.1",
  "apprise>=1.7.0",
  "python-multipart>=0.0.9"
]

[project.scripts]
scriberr-gateway = "scriberr_gateway.server:run"

[tool.setuptools]
packages = ["scriberr_gateway"]
